Add setter methods for group kernel clock multiplexers.
This is the last piece of functionality that was missing from the Peripheral REC
implementation. This implementation requires all peripherals to be
available (through a `&mut`) in order to set group kernel clocks. This ensures
it won't affect any active periperals.

It could be made slightly more flexible by allowing changes when all the
periperals in the group are available, but that would be more complicated.
